requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
cffi==1.11.5
coloredlogs==10.0
cryptography==2.4.2
future==0.17.1
humanfriendly==4.17
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.4.0
pycparser==2.19
pylint==2.2.2
python-dotenv==0.9.1
python-telegram-bot==11.1.0
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
